+++ Krowi_AchievementFilter/Localization/enUS.lua
@@ -133,6 +133,9 @@ L["Skin Ace3"] = "Skin Ace3";
 L["Skin Ace3 Desc"] = "Applies the ElvUI skin to the Options.\n-> Ace3";
 
 -- Options Donations
+L["Credits"] = "Credits";
+L["Special thanks"] = "Special thanks";
+L["Special thanks Desc"] = "Special thanks to {bur} ";
 L["Donations"] = "Donations";
 
 -- Keybindings
@@ -239,16 +242,17 @@ L["Enhanced tooltip Categories Desc"] = "Each tooltip has additional information
                                         "The {red} part of the progress bar and number in brackets can either be disbled in options ({showNotObtainable}) or by unchecking the {notObtainable} filter.";
 L["Enhanced tooltip Achievements Desc"] = "Each tooltip has additional information on how achievements relate to each other (if they do). Several new sections are added to it.\n\n" ..
                                             "Features:\n" ..
-                                            AF_TAB .. "- {partOfAChain}: Shows the list of achievements the hovered achievement is part of ({partOfAChainEnabled})\n" ..
-                                            AF_TAB .. "- {requiredFor}: Shows a list of achievements that require the hovered achievement to be completed in order to earn it ({requiredForEnabled})\n\n" ..
+                                            AF_TAB .. "- {partOfAChain}: Shows the list of achievements the hovered achievement is part of (1) ({partOfAChainEnabled})\n" ..
+                                            AF_TAB .. "- {requiredFor}: Shows a list of achievements that require the hovered achievement to be completed in order to earn it (1) ({requiredForEnabled})\n" ..
+                                            AF_TAB .. "- {objectivesProgress}: Shows the list of objectives with progress ({objectivesProgressEnabled})\n\n" ..
                                             "Icons:\n" ..
                                             AF_TAB .. "- {ready}: The achievement is completed\n" ..
                                             AF_TAB .. "- {waiting}: The achievement is not completed\n" ..
                                             AF_TAB .. "- {notready}: The achievement is not possible to obtain anymore\n\n" ..
-                                            "Optionally, two icons can be displayed per line. The first icon is the default one, the second one is for the current character. ({currentCharacterIconEnabled})\n\n" ..
+                                            "(1) Optionally, two icons can be displayed per line. The first icon is the default one, the second one is for the current character. ({currentCharacterIconEnabled})\n\n" ..
                                             "Colors:\n" ..
-                                            AF_TAB .. "- {green}: The achievement is completed\n" ..
-                                            AF_TAB .. "- {grey}: The achievement is not completed\n" ..
+                                            AF_TAB .. "- {green}: The achievement/objective is completed\n" ..
+                                            AF_TAB .. "- {grey}: The achievement/objective is not completed\n" ..
                                             AF_TAB .. "- {red}: The achievement is not possible to obtain anymore\n" ..
                                             AF_TAB .. "- Lighter {lightGreen}, {lightGrey} or {lightRed}: The hovered achievement";